+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/FVPEventBridge.m
@@ -79,6 +79,15 @@
   }];
 }
 
+- (void)videoPlayerDidLoadUrlWithDuration:(int64_t)duration size:(CGSize)size {
+  [self sendOrQueue:@{
+    @"event" : @"urlLoaded",
+    @"duration" : @(duration),
+    @"width" : @(size.width),
+    @"height" : @(size.height)
+  }];
+}
+
 - (void)videoPlayerDidErrorWithMessage:(NSString *)errorMessage {
   [self sendOrQueue:[FlutterError errorWithCode:@"VideoError" message:errorMessage details:nil]];
 }